+++ src/Driver/Pdo/AbstractPdoConnection.php
@@ -19,7 +19,6 @@ use Yiisoft\Db\Profiler\Context\ConnectionContext;
 use Yiisoft\Db\Profiler\ProfilerAwareInterface;
 use Yiisoft\Db\Profiler\ProfilerAwareTrait;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
-use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 use Yiisoft\Db\Transaction\TransactionInterface;
 
@@ -45,7 +44,6 @@ abstract class AbstractPdoConnection extends AbstractConnection implements PdoCo
     protected string $serverVersion = '';
     protected bool|null $emulatePrepare = null;
     protected QueryBuilderInterface|null $queryBuilder = null;
-    protected QuoterInterface|null $quoter = null;
     protected SchemaInterface|null $schema = null;
 
     public function __construct(protected PdoDriverInterface $driver, protected SchemaCache $schemaCache)